Leave highlighting on when code expanded with hidden matches
It auto-focuses which hides highlighting but we want the user to
see the matches since we are the ones who prompted them to click on
it to see the matches.

The behaviour of the code editor is odd but it works in our favour
since it automatically hides highlighting when clicking on an inner
html part, and somehow it's restored when clickout out of it. Neat.