Tag fixes: handle :: properly and improve match counting
Original assumption about :: was wrong. Assume the original data
(and Anki's search) is always ::, so we only consider :: for matching.
∷ is only considered for displaying, which means we adjust the match
length to account for 1 char instead of 2.

Allow a regex to only count a tag match once per tag. This fixes cases
where the user can do a regex search that may have many matches in one
tag, which is very confusing.