Expand range of css selectors for Windows build, hide color menu
It seems that earlier builds of the browser engine do not permit
us to grab css vars in a ::highlight section that were defined
anywhere but a ::highlight section. This resulted in no highlighting
on that build (which Windows is still on). So this fixes the no
highlighting issue but need to explore ways to get the custom colors
working again. It's going to be JS most likely :(